Just use UInt since we are 64 bit

+++ Sources/TaskSupport/TaskSupport.swift
@@ -15,9 +15,9 @@ public typealias activity_scope_state_s = os_activity_scope_state_s
 
 #else
 
-public typealias activity_id_t = UInt64
-public typealias parent_activity_id_t = UInt64  // there's no way to obtain a context parent on Linux
-public typealias activity_scope_state_s = UInt64  // this is an opaque structure on MacOS
+public typealias activity_id_t = UInt
+public typealias parent_activity_id_t = UInt  // there's no way to obtain a context parent on Linux
+public typealias activity_scope_state_s = UInt  // this is an opaque structure on MacOS
 
 #endif
